relocate interfaces for oauth
