+++ src/MyoFibril.WebAPI/OAuth/AuthHeaderHandler.cs
@@ -1,4 +1,5 @@
 ﻿using System.Net.Http.Headers;
+using MyoFibril.WebAPI.OAuth.Interfaces;
 
 namespace MyoFibril.WebAPI.OAuth;
