+++ src/MyoFibril.WebAPI/OAuth/Interfaces/ITokenCache.cs
@@ -1,4 +1,4 @@
-﻿namespace MyoFibril.WebAPI.OAuth;
+﻿namespace MyoFibril.WebAPI.OAuth.Interfaces;
 public interface ITokenCache
 {
     void SetAccessToken(string accessToken);
