+++ src/MyoFibril.WebAPI/OAuth/TokenRefreshService.cs
@@ -1,4 +1,5 @@
 ﻿using MyoFibril.Contracts.Strava.Responses.OAuth;
+using MyoFibril.WebAPI.OAuth.Interfaces;
 using System.Text.Json;
 using System.Web;
